requirements.txt
requests
beautifulsoup4
pymongo
deep-translator
python-telegram-bot
pytz
docx2pdf
python-dotenv
python-docx
pypandoc
pdfkit
